refactor(cli): create package internal/cli
- move `ansi` -> `internal/cli/ansi`
- move `internal/configflags` -> `internal/cli/configflags`
